requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
email-validator==2.2.0
bcrypt==4.2.1
requests==2.32.3
beautifulsoup4==4.12.3
fpdf2==2.8.2
python-dotenv==1.0.1
waitress==3.0.2
